+++ dessia_common/workflow/core.py
@@ -11,23 +11,24 @@ import json
 import webbrowser
 
 import io
-import networkx as nx
 from typing import List, Union, Type, Any, Dict, Tuple, Optional
 from copy import deepcopy
+import warnings
+import traceback as tb
+import networkx as nx
 from dessia_common.templates import workflow_template
-from dessia_common import DessiaObject, is_sequence,\
-    JSONSCHEMA_HEADER, jsonschema_from_annotation,\
+from dessia_common import DessiaObject, is_sequence, JSONSCHEMA_HEADER, jsonschema_from_annotation,\
     deserialize_argument, set_default_value, prettyname, serialize_dict, DisplaySetting
 
 from dessia_common.utils.serialization import dict_to_object, deserialize, serialize_with_pointers, serialize,\
                                               dereference_jsonpointers
-from dessia_common.utils.types import serialize_typing,\
-    deserialize_typing, recursive_type, typematch
+from dessia_common.utils.types import serialize_typing, deserialize_typing, recursive_type, typematch
 from dessia_common.utils.copy import deepcopy_value
 from dessia_common.utils.docstrings import FAILED_ATTRIBUTE_PARSING, EMPTY_PARSED_ATTRIBUTE
 from dessia_common.utils.diff import choose_hash
 from dessia_common.typings import JsonSerializable, MethodType
-import warnings
+from dessia_common.displays import DisplayObject
+from dessia_common.breakdown import attrmethod_getter
 
 
 class Variable(DessiaObject):
@@ -451,9 +452,8 @@ class Workflow(Block):
         """
         Computes the displays of the objects
         """
-        display_settings = []
-        display_settings.append(DisplaySetting('documentation', 'markdown', 'to_markdown', None))
-        display_settings.append(DisplaySetting('workflow', 'workflow', 'to_dict', None))
+        display_settings = [DisplaySetting('documentation', 'markdown', 'to_markdown', None),
+                            DisplaySetting('workflow', 'workflow', 'to_dict', None)]
         return display_settings
 
     def to_markdown(self):
@@ -1332,7 +1332,7 @@ class WorkflowState(DessiaObject):
                                                              path=f"{path}/input_values/{input_}")
             else:
                 serialized_v = serialize(value)
-            input_values[input_] = serialized_v
+            input_values[str(input_)] = serialized_v
 
         dict_['input_values'] = input_values
 
@@ -1354,9 +1354,9 @@ class WorkflowState(DessiaObject):
                 variable_index = self.workflow.variable_index(variable)
                 serialized_value, memo = serialize_with_pointers(value=value, memo=memo,
                                                                  path=f"{path}/values/{variable_index}")
-                values[variable_index] = serialized_value
+                values[str(variable_index)] = serialized_value
         else:
-            values = {self.workflow.variable_index(i): serialize(v) for i, v in self.values.items()}
+            values = {str(self.workflow.variable_index(i)): serialize(v) for i, v in self.values.items()}
 
         dict_['values'] = values
 
@@ -1397,17 +1397,14 @@ class WorkflowState(DessiaObject):
         values = {}
         if 'values' in dict_:
             for i, value in dict_['values'].items():
-                values[workflow.variables[int(i)]] = deserialize(value,
-                                                                 global_dict=dict_,
+                values[workflow.variables[int(i)]] = deserialize(value, global_dict=dict_,
                                                                  pointers_memo=pointers_memo,
                                                                  path=f'{path}/values/{i}')
-        elif 'variable_values' in dict_:
-            # Retrocompat with variable value may be removed after v0.10.0:
-            for i, value in dict_['variable_values'].items():
-                values[workflow.variables[int(i)]] = deserialize(value,
-                                                                 global_dict=dict_,
-                                                                 pointers_memo=pointers_memo,
-                                                                 path=f'{path}/variable_values/{i}')
+        # elif 'variable_values' in dict_:
+        #     for i, value in dict_['variable_values'].items():
+        #         values[workflow.variables[int(i)]] = deserialize(value, global_dict=dict_,
+        #                                                          pointers_memo=pointers_memo,
+        #                                                          path=f'{path}/variable_values/{i}')
 
         input_values = {int(i): deserialize(v, global_dict=dict_, pointers_memo=pointers_memo)
                         for i, v in dict_['input_values'].items()}
@@ -1469,9 +1466,7 @@ class WorkflowState(DessiaObject):
         """
         Computes the displays of the objects
         """
-        display_settings = []
-        display_settings.append(DisplaySetting('workflow-state', 'workflow_state', 'to_dict', None))
-        return display_settings
+        return [DisplaySetting('workflow-state', 'workflow_state', 'to_dict', None)]
 
     @property
     def progress(self):
@@ -1727,7 +1722,6 @@ class WorkflowRun(WorkflowState):
         """
         Adds variable values to super WorkflowState dict
         """
-
         if memo is None:
             memo = {}  # To make sure we have the good ref for next steps
         dict_ = WorkflowState.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
@@ -1735,14 +1729,16 @@ class WorkflowRun(WorkflowState):
         # To force migrating from dessia_common.workflow
         dict_['object_class'] = 'dessia_common.workflow.core.WorkflowRun'
 
-        if use_pointers:
-            variable_values = {}
-            for key, value in variable_values.items():
-                variable_values[str(key)], memo = serialize_with_pointers(value, memo=memo,
-                                                                          path=f'{path}/variable_values/{key}')
-            dict_["variable_values"] = variable_values
-        else:
-            dict_["variable_values"] = {str(k): serialize(v) for k, v in self.variable_values.items()}
+        # TODO REMOVING THIS TEMPORARLY TO PREVENT DISPLAYS TO BE LOST WITH POINTERS
+        # VARIABLE_VALUES ARE NOT SET BACK IN DICT_TO_OBJECT
+        # if use_pointers:
+        #     variable_values = {}
+        #     for key, value in self.variable_values.items():
+        #         variable_values[str(key)], memo = serialize_with_pointers(value, memo=memo,
+        #                                                                   path=f'{path}/variable_values/{key}')
+        #     dict_["variable_values"] = variable_values
+        # else:
+        dict_["variable_values"] = {str(k): serialize(v) for k, v in self.variable_values.items()}
         return dict_
 
     def display_settings(self) -> List[DisplaySetting]:
@@ -1754,14 +1750,14 @@ class WorkflowRun(WorkflowState):
         display_settings.append(DisplaySetting('workflow-state', 'workflow_state', 'to_dict', None))
 
         # Find & order displayable blocks
-        d_blocks = [b for b in self.workflow.blocks if hasattr(b, 'display_')]
+        d_blocks = [b for b in self.workflow.blocks if hasattr(b, 'display_') and hasattr(b, "_display_settings")]
+        # Change last line to isinstance ?
         sorted_d_blocks = sorted(d_blocks, key=lambda b: b.order)
         for block in sorted_d_blocks:
             block_index = self.workflow.blocks.index(block)
-            block_display = block.display_settings(block_index=block_index)
-            block_display.method = 'block_display'
-            block_display.arguments = {'block_index': block_index}
-            display_settings.append(block_display)
+            settings = block._display_settings(block_index)  # Code intel is not working properly here
+            if settings is not None:
+                display_settings.append(settings)
 
         if isinstance(self.output_value, DessiaObject):
             output_display_settings = [ds.compose('output_value') for ds in self.output_value.display_settings()]
@@ -1769,16 +1765,41 @@ class WorkflowRun(WorkflowState):
 
         return display_settings
 
+    def _display_from_selector(self, selector: str, **kwargs) -> DisplayObject:
+        """
+        Generate the display from the selector
+        """
+        # TODO THIS IS A TEMPORARY DIRTY HOTFIX OVERWRITE.
+        #  WE SHOULD IMPLEMENT A WAY TO GET RID OF REFERENCE PATH WITH URLS
+        track = ""
+        if selector in ["documentation", "workflow"]:
+            return self.workflow._display_from_selector(selector)
+
+        if selector == "workflow-state":
+            return DessiaObject._display_from_selector(self, selector)
+
+        # Displays for blocks (getting reference path from block_display return)
+        display_setting = self._display_settings_from_selector(selector)
+        try:
+            data, reference_path = attrmethod_getter(self, display_setting.method)(**display_setting.arguments)
+        except:
+            data, reference_path = None, ""
+            track = tb.format_exc()
+
+        if display_setting.serialize_data:
+            data = serialize(data)
+        return DisplayObject(type_=display_setting.type, data=data, reference_path=reference_path, traceback=track)
+
     def block_display(self, block_index: int):
         """
         Computes the display of associated block to use integrate it in the workflow run displays
         """
         self._activate_activable_pipes()
         self.activate_inputs()
-        block = self.blocks[block_index]
+        block = self.workflow.blocks[block_index]
         if block in self._activable_blocks():
             self._evaluate_block(block)
-        reference_path = ''
+        reference_path = ""
         local_values = {}
         for i, input_ in enumerate(block.inputs):
             input_adress = self.workflow.variable_indices(input_)
@@ -1786,7 +1807,7 @@ class WorkflowRun(WorkflowState):
             if i == block._displayable_input:
                 reference_path = f'variable_values/{input_adress}'
         display_ = block.display_(local_values=local_values, reference_path=reference_path)
-        return display_
+        return display_, reference_path
 
     def dict_to_arguments(self, dict_: JsonSerializable, method: str):
         if method in self._allowed_methods: